removing test from code
